requirements.txt
flask>=3.0
flask-cors>=4.0
anthropic>=0.40
openai>=1.50
google-generativeai>=0.8
python-dotenv>=1.0
